backend/requirements.txt
flask==3.1.2
flask-cors==5.0.0
pandas==2.3.3
numpy==2.3.3
scikit-learn==1.7.2
xgboost==2.1.3
lightgbm==4.5.0
folium==0.18.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==23.0.0
pytest==8.3.4
joblib==1.4.2
geopy==2.4.1
matplotlib==3.10.0
seaborn==0.13.2
jupyter==1.1.1
notebook==7.3.1
